Rename app/ directory to api/, restructure as nestjs project, and add info/p2p resource
