+++ packages/networking-service/src/api/index.ts
@@ -6,7 +6,7 @@
 import { Logger } from '@nestjs/common';
 import { NestFactory } from '@nestjs/core';
 
-import { AppModule } from './app.module';
+import { AppModule } from './app/app.module';
 
 class App {
     constructor() {
