--- packages/networking-service/src/app/info/info.controller.ts
@@ -1,10 +1,0 @@
-import { Controller, Get } from '@nestjs/common';
-import upnp from '../../upnp';
-
-@Controller('info')
-export class InfoController {
-    @Get()
-    async findAll() {
-        return upnp.info();
-    }
-}
